move: 사이드바 컴포넌트로 분리
