+++ src/app/(pages)/map/_components/KakaoMap.tsx
@@ -5,6 +5,7 @@ import { Map, MapMarker, MarkerClusterer } from "react-kakao-maps-sdk";
 import { Article, GroupedData } from "../../../types/mapTypes/ArticleType";
 import SidePanel from "./SidePanel";
 import { fetchArticles } from "./supabase";
+import SideBar from "./SideBar";
 
 export default function KaKaoMap() {
   const [articles, setArticles] = useState<Article[]>([]);
@@ -12,7 +13,6 @@ export default function KaKaoMap() {
   const [filteredArticles, setFilteredArticles] = useState<Article[]>([]);
   const [isSidePanelOpen, setIsSidePanelOpen] = useState(true);
   const [isLoading, setIsLoading] = useState(false);
-  const [activeFilter, setActiveFilter] = useState<string>("all");
   const [viewMode, setViewMode] = useState<"all" | "selected">("all");
 
   // 처음 로드 시 Supabase에서 articles 가져오기
@@ -30,24 +30,6 @@ export default function KaKaoMap() {
 
     loadArticles();
   }, []);
-
-  // 필터링 함수
-  const filterArticles = (filter: string) => {
-    setActiveFilter(filter);
-    if (filter === "all") {
-      setFilteredArticles(articles);
-    } else if (filter === "원/투룸") {
-      setFilteredArticles(
-        articles.filter((article) => article.building_type === "원룸" || article.building_type === "투룸")
-      );
-    } else if (filter === "주택/빌라") {
-      setFilteredArticles(
-        articles.filter((article) => article.building_type === "주택" || article.building_type === "빌라")
-      );
-    } else {
-      setFilteredArticles(articles.filter((article) => article.building_type === filter));
-    }
-  };
 
   // 같은 위치의 후기들을 그룹화
   const groupedData: GroupedData = useMemo(() => {
@@ -88,52 +70,17 @@ export default function KaKaoMap() {
     setIsSidePanelOpen(true);
   };
 
-  // 사이드 패널 토글 함수 추가
-  const toggleSidePanel = () => {
-    setIsSidePanelOpen(!isSidePanelOpen);
-  };
-
   if (isLoading) return <div>지도 로딩 중...</div>;
 
   return (
     <div className="flex flex-row">
-      {/* 왼쪽 사이드바 */}
-      <div className="flex flex-col gap-3 w-1/12">
-        <button
-          className={`border rounded-md p-1 ${activeFilter === "all" ? "bg-blue-500 text-white" : ""}`}
-          onClick={() => filterArticles("all")}
-        >
-          모든 후기
-        </button>
-        <button
-          className={`border rounded-md p-1 ${activeFilter === "원/투룸" ? "bg-blue-500 text-white" : ""}`}
-          onClick={() => filterArticles("원/투룸")}
-        >
-          원/투룸
-        </button>
-        <button
-          className={`border rounded-md p-1 ${activeFilter === "아파트" ? "bg-blue-500 text-white" : ""}`}
-          onClick={() => filterArticles("아파트")}
-        >
-          아파트
-        </button>
-        <button
-          className={`border rounded-md p-1 ${activeFilter === "주택/빌라" ? "bg-blue-500 text-white" : ""}`}
-          onClick={() => filterArticles("주택/빌라")}
-        >
-          주택/빌라
-        </button>
-        <button
-          className={`border rounded-md p-1 ${activeFilter === "오피스텔" ? "bg-blue-500 text-white" : ""}`}
-          onClick={() => filterArticles("오피스텔")}
-        >
-          오피스텔
-        </button>
-        {/* 토글버튼 */}
-        <button onClick={toggleSidePanel} className="border rounded-md p-1">
-          검색
-        </button>
-      </div>
+      {/* 사이드 바 */}
+      <SideBar
+        articles={articles}
+        setFilteredArticles={setFilteredArticles}
+        setIsSidePanelOpen={setIsSidePanelOpen}
+        isSidePanelOpen={isSidePanelOpen}
+      />
 
       {/* 지도영역 */}
       <div className="w-11/12 overflow-hidden relative">
